Extracts the delivery extensions to separate classes
