+++ src/PushNotifications.Delivery.FireBase/PushNotifications.Delivery.FireBase.csproj
@@ -47,6 +47,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="App_Packages\LibLog.4.2\LibLog.cs" />
+    <Compile Include="Extensions\FireBaseExtensions.cs" />
     <Compile Include="InternalsVisibleToTests.cs" />
     <Compile Include="FireBaseDelivery.cs" />
     <Compile Include="FireBaseTopicSubscriptionManager.cs" />
